+++ R/break.R
@@ -26,6 +26,12 @@ broken <- function(model, new_observation, ...) {
 #' new_observation <- iris[1,]
 #' br <- broken(model, new_observation)
 #' plot(br)
+#'
+#' # works for interactions as well
+#' model <- lm(Sepal.Length~Petal.Width*Species, data=iris)
+#' new_observation <- iris[1,]
+#' br <- broken(model, new_observation)
+#' plot(br)
 
 broken.lm <- function(model, new_observation, ..., baseline = 0) {
   ny <- predict.lm(model, newdata = new_observation, type="terms")